requirements.txt
altair==5.3.0
anyio==4.4.0
astor==0.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
easyocr==1.7.1
exceptiongroup==1.2.2
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib_resources==6.4.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.5.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.0
paddlepaddle==2.6.1
pandas==2.0.3
pdf2image==1.17.0
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==5.27.2
pyarrow==16.1.0
pyclipper==1.3.0.post5
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
rapidfuzz==3.9.4
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
streamlit==1.36.0
sympy==1.13.0
tenacity==8.5.0
termcolor==2.4.0
tifffile==2023.7.10
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zipp==3.19.2
